backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.2.0
langchain-community==0.2.0
faiss-cpu
langchain-groq
langchain-huggingface
sentence-transformers
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
httpx==0.27.0
python-jose==3.3.0
requests==2.31.0
elevenlabs==1.2.0
certifi
slowapi==0.1.9
stripe
